WIP: Parsing Server world data - this doesnt work atm, because ChunkColumnPtr column = m_world->GetChunkColumn(0, 0); is nullptr...

+++ include/recraft/blocks/CT_Block.h
@@ -8,7 +8,7 @@
 #include "mclib/block/Block.h"
 #include "mclib/inventory/Slot.h"
 
-typedef uint8_t Block;
+typedef u16 Block;
 
 //TODO: Instead of ENUM IDs, use the "flattened" strings
 enum { Block_Air,
